requirements.txt
Flask
mysql-connector-python
flask_httpauth
werkzeug